fix(convar): Remove type restriction in convar generic type
Having a restriction of 'struct' was causing us to be not able to create a convar with a string type. Crazy that I originally made this code ~7 years ago and I've only just realised this was an issue.

Also had to make hacky way of converting convar to enum now, thank you cfx and your incredibly old .net